Move ActionFailed under Raise
